Cloak Sleeves and Cloak Sleeves Feather Cover introduction!
This took a lot of time, but I believe it will be worth it! Have fun!
No, it is not an april fools trick, uploaded in the early hours of the 2nd.

+ Added Scientist Cloak variants
+ Added Seeker's Cloak variants
+ Added Nomad's Cloak variants
+ Added Stalker's Cloak variants

Special thanks to Mobiteq for making this possible, by helping out with
stalker's cloak. https://steamcommunity.com/id/Mobiteq

Signed-off-by: Eradem